Is not an Integer but Long

+++ src/main/java/dk/dataforsyningen/arkivmeta/dokument/apimapper/DokumentDtoMapper.java
@@ -59,9 +59,9 @@ public class DokumentDtoMapper implements RowMapper<DokumentDto> {
         dto.setFiltype(rs.getString("filtype"));
         dto.setDokumentsamling(rs.getString("dokumentsamling"));
         dto.setHerredsnavn(rs.getString("herredsnavn"));
-        dto.setHerredsnummer((Integer) rs.getObject("herredsnummer"));
+        dto.setHerredsnummer((Long) rs.getObject("herredsnummer"));
         dto.setProtokoltype(rs.getString("protokoltype"));
-        dto.setSogneid((Integer) rs.getObject("sogneid"));
+        dto.setSogneid((Long) rs.getObject("sogneid"));
         dto.setSognenavn(rs.getString("sognenavn"));
 
         return dto;
